unk8sify pachctl config tests
