+++ src/server/config/cmds_test.go
@@ -1,5 +1,3 @@
-//go:build k8s
-
 package cmds
 
 import (
@@ -33,10 +31,6 @@ func run(t *testing.T, cmd string) error {
 }
 
 func TestInvalidEnvValue(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
 	require.YesError(t, run(t, `
 		export PACH_CONTEXT=foobar
 		pachctl config get active-context
@@ -44,10 +38,6 @@ func TestInvalidEnvValue(t *testing.T) {
 }
 
 func TestEnvValue(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
 	require.NoError(t, run(t, `
 		echo '{}' | pachctl config set context foo --overwrite
 		echo '{}' | pachctl config set context bar --overwrite
@@ -84,10 +74,6 @@ func TestConnectExistingWithAlias(t *testing.T) {
 }
 
 func TestMetrics(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
 	require.NoError(t, run(t, `
 		pachctl config get metrics | match true
 		pachctl config set metrics false
@@ -98,10 +84,6 @@ func TestMetrics(t *testing.T) {
 }
 
 func TestActiveContext(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
 	require.YesError(t, run(t, `
 		pachctl config set active-context foo 2>&1 | match "context does not exist: foo"
 	`))
@@ -114,10 +96,6 @@ func TestActiveContext(t *testing.T) {
 }
 
 func TestActiveEnterpriseContext(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
 	require.YesError(t, run(t, `
 		pachctl config set active-enterprise-context foo 2>&1 | match "context does not exist: foo"
 	`))
@@ -130,10 +108,6 @@ func TestActiveEnterpriseContext(t *testing.T) {
 }
 
 func TestSetContext(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
 	require.YesError(t, run(t, `
 		echo 'malformed_json' | pachctl config set context foo
 	`))
@@ -151,10 +125,6 @@ func TestSetContext(t *testing.T) {
 }
 
 func TestUpdateContext(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
 	require.YesError(t, run(t, `
 		pachctl config update context foo --pachd-address=bar
 	`))
@@ -174,10 +144,6 @@ func TestUpdateContext(t *testing.T) {
 }
 
 func TestDeleteContext(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
 	require.YesError(t, run(t, `
 		pachctl config delete context foo
 	`))
@@ -189,9 +155,6 @@ func TestDeleteContext(t *testing.T) {
 }
 
 func TestConfigListContext(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
 	// Verify the * marker exists for the active-context when enterprise is disabled and the enterprise context isn't set
 	require.NoError(t, run(t, `
 		echo '{}' | pachctl config set context foo
@@ -211,23 +174,5 @@ func TestConfigListContext(t *testing.T) {
 		pachctl config set active-context foo
 		pachctl config list context | match "	bar"
 		pachctl config list context | match "E\*	foo"
-	`))
-}
-
-func TestImportKube(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
-	require.NoError(t, run(t, `
-		pachctl config import-kube imported
-		pachctl config get active-context | match 'imported'
-		pachctl config get context imported | match "\"cluster_name\":[[:space:]]*\"$(kubectl config current-context)\""
-		pachctl config get context imported | match '"namespace":[[:space:]]*"default"'
-		pachctl config import-kube enterprise-kube --overwrite --namespace enterprise --enterprise
-		pachctl config get active-enterprise-context | match 'enterprise-kube'
-		pachctl config get context enterprise-kube | match "\"cluster_name\":[[:space:]]*\"$(kubectl config current-context)\""
-		pachctl config get context enterprise-kube | match '"namespace":[[:space:]]*"enterprise"'
-
 	`))
 }
